Handle Private Person and Company Registration in Signup Form
This commit includes changes to handle the registration process for companies. Our backend API requires a firstName and lastName for user registration, which does not align with the information companies provide. To solve this, we made modifications to the handleSubmitSignup function in our Authentication page.

The changes made are as follows:

We added companyName to the object destructuring of values obtained from the signup form.
We then checked whether companyName exists to determine if the user is a company or not.
If a company is registering, we treated companyName as firstName, and set lastName as 'null'.
If it's an individual user, we continued using fname and lname as before, trimming any leading or trailing white spaces.
This way, we're now able to handle both individual user registrations and company registrations, without breaking our existing signup flow or making substantial changes to our backend.